refactor(frontend): separate utils into smaller files
