+++ packages/frontend/src/utils/get-formatted-date.ts
@@ -1,0 +1,10 @@
+export const getFormattedDate = (date: string): string => {
+  const dateObj = new Date(date)
+  if (!date || !dateObj) {
+    return ''
+  }
+  const year = dateObj.getFullYear()
+  const month = `${dateObj.getMonth() + 1}`.padStart(2, '0')
+  const day = `${dateObj.getDate()}`.padStart(2, '0')
+  return [year, month, day].join('.')
+}
