+++ packages/frontend/src/utils/get-post-summaries.ts
@@ -1,0 +1,22 @@
+import { PostSummary } from '@/components/types'
+import { DEFAULT_THEME_COLOR, FALLBACK_IMG } from '@/constants'
+
+export const getPostSummaries = (posts: any[]): PostSummary[] => {
+  // TODO: error handling for post
+  return posts?.map((post: any) => {
+    const subSubcategory = post?.subSubcategoriesOrdered?.[0]
+
+    return {
+      image: post?.heroImage?.resized?.small ?? FALLBACK_IMG,
+      title: post.title,
+      url: `/article/${post.slug}`,
+      desc: post.ogDescription,
+      category: subSubcategory?.subcategory?.name,
+      subSubcategory: subSubcategory?.name,
+      publishedDate: post.publishedDate,
+      theme:
+        subSubcategory?.subcategory?.category?.themeColor ||
+        DEFAULT_THEME_COLOR,
+    }
+  })
+}
