+++ packages/frontend/src/utils/get-theme-color.ts
@@ -1,0 +1,11 @@
+import { Theme, ThemeColor } from '@/constants'
+
+export const getThemeColor = (theme: Theme) => {
+  if (theme === Theme.YELLOW) {
+    return ThemeColor.YELLOW
+  } else if (theme === Theme.RED) {
+    return ThemeColor.RED
+  } else {
+    return ThemeColor.BLUE
+  }
+}
