+++ packages/frontend/src/utils/index.ts
@@ -1,136 +1,5 @@
-import errors from '@twreporter/errors'
-import axios, { AxiosRequestConfig } from 'axios'
-
-import { PostSummary } from '@/components/types'
-import {
-  API_URL,
-  DEFAULT_THEME_COLOR,
-  FALLBACK_IMG,
-  INTERNAL_API_URL,
-  Theme,
-  ThemeColor,
-} from '@/constants'
-import envVars from '@/environment-variables'
-
-export const getThemeColor = (theme: Theme) => {
-  if (theme === Theme.YELLOW) {
-    return ThemeColor.YELLOW
-  } else if (theme === Theme.RED) {
-    return ThemeColor.RED
-  } else {
-    return ThemeColor.BLUE
-  }
-}
-
-export const getFormattedDate = (date: string): string => {
-  const dateObj = new Date(date)
-  if (!date || !dateObj) {
-    return ''
-  }
-  const year = dateObj.getFullYear()
-  const month = `${dateObj.getMonth() + 1}`.padStart(2, '0')
-  const day = `${dateObj.getDate()}`.padStart(2, '0')
-  return [year, month, day].join('.')
-}
-
-export const getPostSummaries = (posts: any[]): PostSummary[] => {
-  // TODO: error handling for post
-  return posts?.map((post: any) => {
-    const subSubcategory = post?.subSubcategoriesOrdered?.[0]
-
-    return {
-      image: post?.heroImage?.resized?.small ?? FALLBACK_IMG,
-      title: post.title,
-      url: `/article/${post.slug}`,
-      desc: post.ogDescription,
-      category: subSubcategory?.subcategory?.name,
-      subSubcategory: subSubcategory?.name,
-      publishedDate: post.publishedDate,
-      theme:
-        subSubcategory?.subcategory?.category?.themeColor ||
-        DEFAULT_THEME_COLOR,
-    }
-  })
-}
-
-export const AXIOS_TIMEOUT = 5000
-
-export const sendGQLRequest = async (
-  data?: any,
-  config?: AxiosRequestConfig<any> | undefined
-) => {
-  // Define url based on environment, dev/staging needs pure internal api to bypass Identity-Aware Proxy(IAP)
-  let url
-  if (typeof window === 'undefined' && !envVars.isProduction) {
-    url = INTERNAL_API_URL
-  } else {
-    url = API_URL
-  }
-
-  let response
-  try {
-    response = await axios.post(url, data, {
-      timeout: AXIOS_TIMEOUT,
-      ...config,
-    })
-  } catch (err) {
-    const annotatedErr = errors.helpers.annotateAxiosError(err)
-    const msg = errors.helpers.printAll(annotatedErr, {
-      withStack: true,
-      withPayload: true,
-    })
-    log(LogLevel.ERROR, msg)
-  }
-
-  const gqlErrors = response?.data?.errors
-  if (gqlErrors) {
-    const annotatedErr = errors.helpers.wrap(
-      new Error(`Errors occured while executing GQL query: ${data}`),
-      'GraphQLError',
-      'Errors occured after axios request',
-      { errors: gqlErrors }
-    )
-    const msg = errors.helpers.printAll(annotatedErr, {
-      withStack: true,
-      withPayload: true,
-    })
-    log(LogLevel.ERROR, msg)
-  }
-
-  return response
-}
-
-export enum LogLevel {
-  INFO = 'INFO',
-  WARNING = 'WARNING',
-  ERROR = 'ERROR',
-}
-
-export const log = (level: LogLevel = LogLevel.INFO, msg: string) => {
-  const structuredMsg = JSON.stringify({
-    severity: level,
-    message: msg,
-  })
-
-  switch (level) {
-    case LogLevel.ERROR: {
-      // Follow https://cloud.google.com/error-reporting/docs/formatting-error-messages doc to print structured error log
-      // and trigger GCP error reporting.
-      const errorLogEntry = {
-        severity: level,
-        '@type':
-          'type.googleapis.com/google.devtools.clouderrorreporting.v1beta1.ReportedErrorEvent',
-        message: msg,
-      }
-      console.error(JSON.stringify(errorLogEntry))
-      return
-    }
-    case LogLevel.WARNING:
-      console.warn(structuredMsg)
-      return
-    case LogLevel.INFO:
-    default:
-      console.log(structuredMsg)
-      return
-  }
-}
+export * from './get-formatted-date'
+export * from './get-post-summaries'
+export * from './get-theme-color'
+export * from './log'
+export * from './send-gql-request'
